feat(nap-serv): extract shared spreadsheet helpers and add import/export to all source entity models
Refactor the inline export/import logic from Employees into a config-driven
spreadsheetHelpers module, then wire up Clients, Contacts, InterCompanies,
and Vendors with the same multi-sheet upsert import/export capability.
