+++ apps/nap-serv/src/system/core/models/InterCompanies.js
@@ -1,5 +1,5 @@
 /**
- * @file InterCompanies model — extends TableModel for tenant-scope inter-companies
+ * @file InterCompanies model — extends TableModel with multi-sheet upsert export/import
  * @module core/models/InterCompanies
  *
  * Copyright (c) 2025 – present NapSoft LLC. All rights reserved.
@@ -7,9 +7,48 @@
 
 import { TableModel } from 'pg-schemata';
 import interCompaniesSchema from '../schemas/interCompaniesSchema.js';
+import {
+  exportSourceEntity,
+  importSourceEntity,
+  PHONE_HEADERS,
+  ADDRESS_HEADERS,
+  TAX_ID_HEADERS,
+} from '../../../lib/spreadsheetHelpers.js';
+
+/** @type {import('../../../lib/spreadsheetHelpers.js').SourceEntityConfig} */
+const CONFIG = {
+  entityName: 'inter_companies',
+  sheetName: 'Inter-Companies',
+  sourceType: 'inter_company',
+  linkColName: 'inter_company_id',
+  idType: null,
+  buildLabel: (row) => row.name,
+  returningCols: ['id', 'name'],
+  boolCols: [],
+  hasRoles: false,
+  codeRequired: true,
+  extraExportCols: [
+    { name: 'status', derive: (origRow) => (origRow.deactivated_at ? 'archived' : 'active') },
+  ],
+  extraImportStrip: [],
+  childSheets: [
+    { sheetName: 'Phone Numbers', modelName: 'phoneNumbers', headers: PHONE_HEADERS },
+    { sheetName: 'Addresses', modelName: 'addresses', headers: ADDRESS_HEADERS },
+    { sheetName: 'Tax Identifiers', modelName: 'taxIdentifiers', headers: TAX_ID_HEADERS },
+  ],
+  appUserProvisioning: null,
+};
 
 export default class InterCompanies extends TableModel {
   constructor(db, pgp, logger = null) {
     super(db, pgp, interCompaniesSchema, logger);
+  }
+
+  async exportToSpreadsheet(filePath, where = [], joinType = 'AND', options = {}) {
+    return exportSourceEntity(this, filePath, where, joinType, options, CONFIG);
+  }
+
+  async importFromSpreadsheet(filePath, _sheetIndex = 0, callbackFn = null, _returning = null) {
+    return importSourceEntity(this, filePath, _sheetIndex, callbackFn, CONFIG);
   }
 }
